調整sql和新增產品統計頁面

+++ CompanyDemo.Database/dbo/Table/ProductT.sql
@@ -3,7 +3,7 @@ CREATE TABLE ProductT(
 	ProductID int NOT NULL IDENTITY,
 	ProductName varchar(100) NOT NULL,
 	ProductType varchar(50) NOT NULL,
-	Price decimal(2,2) NOT NULL,
+	Price INT NOT NULL,
 	Unit varchar(10) NULL,
 	CreatedDate datetime NOT NULL DEFAULT GETUTCDATE(),
 	EditedDate datetime NOT NULL DEFAULT GETUTCDATE(),
